Move assert_region_allclose to test helpers
